[Refactor] handleError 위치 변경, GptSavedList 네브바 수정
